% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadTree.R
\name{filterByDistance}
\alias{filterByDistance}
\title{Helper to verify exact distance after grid pruning}
\usage{
filterByDistance(
  queryPts,
  worldPts,
  candidatesList,
  radius,
  queryRadii,
  worldRadii,
  useWeighted
)
}
\arguments{
\item{queryPts}{Matrix of query points.}

\item{worldPts}{Matrix of world points.}

\item{candidatesList}{List of candidate indices.}

\item{radius}{Search radius.}

\item{queryRadii}{Radii for query points.}

\item{worldRadii}{Radii for world points.}

\item{useWeighted}{Logical flag to use weighted distances.}
}
\value{
List of vectors (list of lists containing valid indices and distances).
}
\description{
Helper to verify exact distance after grid pruning. Returns indices of World points within radius of Query points.
}
