% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadTree.R
\name{gridIndexCreate}
\alias{gridIndexCreate}
\title{Create a Grid Index from a Matrix of coordinates}
\usage{
gridIndexCreate(pts, cellSize)
}
\arguments{
\item{pts}{Matrix (N x 2) with X, Y.}

\item{cellSize}{Size of grid bins.}
}
\value{
An object of class "GridIndex" containing the index, cellSize, and pts.
}
\description{
Create a Grid Index from a Matrix of coordinates. Implements a Spatial Grid Index for fast neighborhood search.
}
