% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadTree.R
\name{gridIndexQueryBatch}
\alias{gridIndexQueryBatch}
\title{Query the Grid for candidates near multiple query points}
\usage{
gridIndexQueryBatch(grid, queryPts, radius)
}
\arguments{
\item{grid}{The GridIndex object.}

\item{queryPts}{Matrix (M x 2).}

\item{radius}{Search radius.}
}
\value{
List of integer vectors (indices in the original pts matrix).
}
\description{
Query the Grid for candidates near multiple query points.
}
