% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{homogeneous2DSetTrafoFromVectors}
\alias{homogeneous2DSetTrafoFromVectors}
\title{Estimate 2D Rigid Body Transformation}
\usage{
homogeneous2DSetTrafoFromVectors(X1, Y1, X2, Y2)
}
\arguments{
\item{X1}{Local points (Source) X coordinates.}

\item{Y1}{Local points (Source) Y coordinates.}

\item{X2}{World points (Target) X coordinates.}

\item{Y2}{World points (Target) Y coordinates.}
}
\value{
A 4x4 transformation matrix.
}
\description{
Estimate 2D Rigid Body Transformation (SVD method).
}
