% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{homogeneousTrafoMatrix2D}
\alias{homogeneousTrafoMatrix2D}
\title{Apply 2D transform to a matrix of points}
\usage{
homogeneousTrafoMatrix2D(M, pts)
}
\arguments{
\item{M}{The transformation matrix.}

\item{pts}{Matrix with columns X, Y.}
}
\value{
Returns N x 2 matrix.
}
\description{
Apply 2D transform to a matrix of points (N x 2 or N x 3).
}
