% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runDemoPlotMatch}
\alias{runDemoPlotMatch}
\title{Run Plot Match Demo}
\usage{
runDemoPlotMatch()
}
\value{
None. Outputs statistics to the console and saves transformed tables to a temporary directory.
}
\description{
Runs a demonstration of the plot matching process using sample data (Air_Positions.txt and Ground_Positions.txt).
}
