% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mls_tools.R
\name{tableCalculateLinkStats}
\alias{tableCalculateLinkStats}
\title{Calculate Link Stats}
\usage{
tableCalculateLinkStats(df, lblX1, lblY1, lblX2, lblY2)
}
\arguments{
\item{df}{Data frame containing linked points.}

\item{lblX1}{Column name for the X coordinate of the first set.}

\item{lblY1}{Column name for the Y coordinate of the first set.}

\item{lblX2}{Column name for the X coordinate of the second set.}

\item{lblY2}{Column name for the Y coordinate of the second set.}
}
\value{
A list containing the mean, standard deviation, minimum, and maximum of the distances.
}
\description{
Calculate distance statistics for linked coordinates.
}
