% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{TableHeaderLabels}
\alias{TableHeaderLabels}
\title{Helper class replacement (List)}
\usage{
TableHeaderLabels(id, x, y, z, r, t, scaleR, scan)
}
\arguments{
\item{id}{Column name for ID, or list of IDs.}

\item{x}{Column name for X.}

\item{y}{Column name for Y.}

\item{z}{Column name for Z, or NULL.}

\item{r}{Column name for Radius, or NULL.}

\item{t}{Column name for Time.}

\item{scaleR}{Scaling factor for radius.}

\item{scan}{Column name for Scan, or NULL.}
}
\value{
A list containing the mapped metadata column labels and scales.
}
\description{
Meta Data Definitions. Helper class replacement (List) to define table header labels.
}
