% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mls_tools.R
\name{tableMatchAndGetLinkTableAndTrafoMatLocal}
\alias{tableMatchAndGetLinkTableAndTrafoMatLocal}
\title{Wrapper for Matcher}
\usage{
tableMatchAndGetLinkTableAndTrafoMatLocal(
  tbWorld,
  tbLocal,
  metaWorld,
  metaLocal,
  searchDist,
  estX0,
  estY0
)
}
\arguments{
\item{tbWorld}{Data frame of the world table.}

\item{tbLocal}{Data frame of the local table.}

\item{metaWorld}{Metadata list for the world table.}

\item{metaLocal}{Metadata list for the local table.}

\item{searchDist}{Maximum search distance.}

\item{estX0}{Estimated X coordinate for the center in world space.}

\item{estY0}{Estimated Y coordinate for the center in world space.}
}
\value{
A list containing the linked table ("tbLink") and the transformation matrix (M).
}
\description{
Wrapper to run the matcher, get the link table, and compute the local transformation matrix.
}
