% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mls_tools.R
\name{tableRead}
\alias{tableRead}
\title{Read a table}
\usage{
tableRead(tablePath)
}
\arguments{
\item{tablePath}{Path to the table file.}
}
\value{
A data frame containing the table data.
}
\description{
Read a table from a file path.
}
