% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mls_tools.R
\name{tableSave}
\alias{tableSave}
\title{Save a table}
\usage{
tableSave(df, outPath)
}
\arguments{
\item{df}{The data frame to save.}

\item{outPath}{The output file path.}
}
\value{
None.
}
\description{
Save a data frame to a file path.
}
