% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mls_tools.R
\name{tableToMatrix}
\alias{tableToMatrix}
\title{Convert table to matrix}
\usage{
tableToMatrix(df, meta)
}
\arguments{
\item{df}{The data frame to convert.}

\item{meta}{Metadata list containing column labels and scale settings.}
}
\value{
A list containing IDs, the extracted matrix, and the original data frame.
}
\description{
Data Preparation. Converts a data frame to a matrix based on metadata.
}
