% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{trafoMatrixCreateAsUnit}
\alias{trafoMatrixCreateAsUnit}
\title{Create a 4x4 Identity Matrix}
\usage{
trafoMatrixCreateAsUnit()
}
\value{
A 4x4 identity matrix.
}
\description{
Create a 4x4 Identity Matrix.
}
